% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_model.R
\name{docformer_pretrain}
\alias{docformer_pretrain}
\title{Docformer Self-supervised training module}
\usage{
docformer_pretrain(..., mask_id = 0)
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[=docformer_config]{docformer_config()}}.}
}
\value{
A \code{luz_module} that has been setup and is ready to be \code{fitted}.
}
\description{
Docformer Self-supervised training module
}
\section{Functions}{
\itemize{
\item \code{docformer_pretrain()}: Pretraining module

}}
\seealso{
\code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}} for fit arguments. See
\code{\link[=predict.docformer_result]{predict.docformer_result()}} for information on how to generate predictions.
}
