% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_model.R
\name{docformer_pretrain}
\alias{docformer_pretrain}
\title{Docformer Self-supervised training task}
\usage{
docformer_pretrain(
  config,
  train_dataloader,
  val_dataloader,
  device,
  epochs,
  path,
  classes,
  lr = 5e-05,
  weights = weights
)
}
\arguments{
\item{config}{config}

\item{train_dataloader}{training dataloader to use}

\item{val_dataloader}{validation dataloader to use}

\item{device}{"cpu" or "cuda" default to "auto"}

\item{epochs}{number of epoch to train}

\item{path}{path}

\item{classes}{number of classes}

\item{lr}{learning-rate}

\item{weights}{weights}
}
\value{
a docformer
}
\description{
Docformer Self-supervised training task
}
