% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.load_weights}
\alias{.load_weights}
\title{Load Pretrained Weights into a Transformers Model}
\usage{
.load_weights(
  model,
  model_name = "microsoft/layoutlm-base-uncased",
  redownload = FALSE,
  timeout = 720
)
}
\arguments{
\item{model}{A transformers-type \code{nn_module} model.}

\item{model_name}{Character; which public Transformers model weights to use. Must be compatible
with \code{model} architecture!. Can be an local file name.}

\item{redownload}{Logical: Shall we force redownload the model weights ?}
}
\value{
The number of model parameters updated. (This is to enable error
checks; the function is called for side effects.)
}
\description{
Loads specified pretrained weights into the given BERT model.
}
\keyword{internal}
