% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{.tokenize}
\alias{.tokenize}
\title{tokenize the character vector and prepend the [CLS] token to first}
\usage{
.tokenize(tokenizer, ...)
}
\arguments{
\item{tokenizer}{the tokenizer function}

\item{...}{unused}

\item{x}{character vector to encode}
}
\value{
list of token ids for each token
}
\description{
tokenize the character vector and prepend the [CLS] token to first
}
