% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_cache}
\alias{download_and_cache}
\title{Download and Cache Weights (the torchvision way)}
\usage{
download_and_cache(url, redownload = FALSE, timeout = 720)
}
\arguments{
\item{url}{the URL of the model to download}

\item{redownload}{should the weights be downloaded fresh even if
they're cached? This is not currently exposed to the end user, and exists}

\item{timeout}{the download timeout}
}
\value{
path of the cached file
}
\description{
Download and Cache Weights (the torchvision way)
}
\examples{
url <- "https://media.githubusercontent.com/media/cregouby/docformer_models/main/inst/tiny-layoutlm.pth"
weight <- download_and_cache(url=url)
}
