% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMEmbeddings}
\alias{LayoutLMEmbeddings}
\title{Construct the embeddings from word, position and token_type embeddings.}
\usage{
LayoutLMEmbeddings(config)
}
\description{
Construct the embeddings from word, position and token_type embeddings.
}
