% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMForMaskedLM}
\alias{LayoutLMForMaskedLM}
\title{LayoutLM Model with a \verb{language modeling} head on top."}
\usage{
LayoutLMForMaskedLM(config)
}
\description{
LayoutLM Model with a \verb{language modeling} head on top."
}
