% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMForSequenceClassification}
\alias{LayoutLMForSequenceClassification}
\title{LayoutLM Model with a sequence classification head on top (a linear layer on top of the pooled output) e.g. for
document image classification tasks such as the \href{https://www.cs.cmu.edu/~aharley/rvl-cdip/}{RVL-CDIP} dataset.}
\usage{
LayoutLMForSequenceClassification(config)
}
\description{
LayoutLM Model with a sequence classification head on top (a linear layer on top of the pooled output) e.g. for
document image classification tasks such as the \href{https://www.cs.cmu.edu/~aharley/rvl-cdip/}{RVL-CDIP} dataset.
}
