% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMForTokenClassification}
\alias{LayoutLMForTokenClassification}
\title{LayoutLM Model with a token classification head on top (a linear layer on top of the hidden-states output) e.g. for
sequence labeling (information extraction) tasks such as the \href{https://guillaumejaume.github.io/FUNSD/}{FUNSD}
dataset and the \href{https://rrc.cvc.uab.es/?ch=13}{SROIE} dataset.}
\usage{
LayoutLMForTokenClassification(config)
}
\description{
LayoutLM Model with a token classification head on top (a linear layer on top of the hidden-states output) e.g. for
sequence labeling (information extraction) tasks such as the \href{https://guillaumejaume.github.io/FUNSD/}{FUNSD}
dataset and the \href{https://rrc.cvc.uab.es/?ch=13}{SROIE} dataset.
}
