% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMModel}
\alias{LayoutLMModel}
\title{The LayoutLM model}
\usage{
LayoutLMModel(config)
}
\arguments{
\item{config}{list of all the parameters of the model.
Initializing with a config file does not load the weights associated with the model, only the
configuration}
}
\value{
a torch::nn_module() of class BaseModelOutputWithPoolingAndCrossAttentions
}
\description{
The LayoutLM model was proposed in \href{https://arxiv.org/abs/1912.13318}{LayoutLM: Pre-training of Text and Layout for Document Image Understanding}
by Yiheng Xu, Minghao Li, Lei Cui, Shaohan Huang, Furu Wei and Ming Zhou.
}
\details{
This model is a torch \href{https://torch.mlverse.org/docs/reference/nn_module.html}{nn_module()}. Use
it as a regular module and refer to the documentation for all matter related to general usage and
behavior.
}
