% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMPreTrainedModel}
\alias{LayoutLMPreTrainedModel}
\title{An abstract class to handle weights initialization and a simple interface for downloading and loading pretrained models.}
\usage{
LayoutLMPreTrainedModel(PreTrainedModel)
}
\description{
An abstract class to handle weights initialization and a simple interface for downloading and loading pretrained models.
}
