% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{normalize_box}
\alias{normalize_box}
\title{Normalize a bounding-box}
\usage{
normalize_box(bbox, width, height, size = 1000)
}
\arguments{
\item{width}{(int): total width of the image}

\item{height}{(int): total height of the image}

\item{size}{(int): target value to normalize bounding_box to. (default 1000)}

\item{box}{(vector of 4 int): the original bounding-box coordinates (xmin, ymin, xmax, ymax)}
}
\value{
a vector of size 4 integers with values normalised to \code{size}.
}
\description{
Takes one or more bounding box and normalize each of their dimensions to \code{size}. If you notice it is
just like calculating percentage except takes \code{size = 1000} instead of 100.
}
\examples{
# normalise bounding-box in percent
normalize_box(c(227,34,274,41), width = 2100, height = 3970, size = 100)
}
