% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{read_featureRDS}
\alias{read_featureRDS}
\title{Load feature tensor from disk}
\usage{
read_featureRDS(file)
}
\arguments{
\item{file}{: source file}
}
\value{

}
\description{
Load feature tensor from disk
}
