% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{save_featureRDS}
\alias{save_featureRDS}
\title{Save feature tensor to disk}
\usage{
save_featureRDS(encoding_lst, file)
}
\arguments{
\item{encoding_lst}{: the feature tensor list to save}

\item{file}{: destination file}
}
\description{
Save feature tensor to disk
}
