% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{.mask_id}
\alias{.mask_id}
\alias{.pad_id}
\alias{.sep_id}
\alias{.cls_id}
\title{Extract special tokens from tokenizer}
\usage{
.mask_id(tokenizer)

.pad_id(tokenizer)

.sep_id(tokenizer)

.cls_id(tokenizer)
}
\description{
Extract special tokens from tokenizer
}
