% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{transformers_config}
\alias{transformers_config}
\title{#' Transformers models configuration}
\format{
A tibble with 5 models observations
}
\source{
\url{https://huggingface.co/} and  \url{https://storage.googleapis.com/torchtransformers-models/}
}
\usage{
transformers_config
}
\description{
A tibble containing the docformers and friends model parameters required to load
model weights from public repositories
}
\keyword{datasets}
