% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{apply_ocr}
\alias{apply_ocr}
\title{Apply tesseract::ocr_data() and clean result.}
\usage{
apply_ocr(image)
}
\arguments{
\item{image}{file path, url, or raw vector to image (png, tiff, jpeg, etc)}
}
\value{
a data.frame of words and associated bounding-box
}
\description{
Apply tesseract::ocr_data() and clean result.
}
\examples{
# good quality scan
df <- apply_ocr(image1)
# poor quality scan
df <- image2 \%>\% image_read() \%>\%
   image_resize("2000x") \%>\%
   image_trim(fuzz = 40) \%>\%
   image_write(format = 'png', density = "300x300") \%>\%
   apply_ocr()
}
