% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{create_features}
\alias{create_features}
\title{Turn image into docformer torch tensor input feature}
\usage{
create_features(
  image,
  tokenizer,
  add_batch_dim = TRUE,
  target_geometry = "224x224",
  max_seq_len = 512,
  save_to_disk = FALSE,
  path_to_save = NULL,
  apply_mask_for_mlm = FALSE,
  extras_for_debugging = FALSE
)
}
\arguments{
\item{image}{file path, url, or raw vector to image (png, tiff, jpeg, etc)}

\item{tokenizer}{tokenizer function to apply to words extracted from image. Currently,
{hftokenizers}, {tokenizer.bpe} and {sentencepiece} tokenizer are supported.}

\item{add_batch_dim}{(boolean) add a extra dimension to tensor for batch encoding}

\item{target_geometry}{image target magik geometry expected by the image model input}

\item{max_seq_len}{size of the embedding vector in tokens}

\item{save_to_disk}{(boolean) shall we save the result onto disk}

\item{path_to_save}{result path}

\item{apply_mask_for_mlm}{add mask to the language model}

\item{extras_for_debugging}{additionnal feature for debugging purposes}
}
\value{

}
\description{
Turn image into docformer torch tensor input feature
}
