% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{create_features_from_doc}
\alias{create_features_from_doc}
\title{Turn document into docformer torch tensor input feature}
\usage{
create_features_from_doc(
  doc,
  tokenizer,
  add_batch_dim = TRUE,
  target_geometry = "500x384",
  max_seq_len = 512,
  save_to_disk = FALSE,
  path_to_save = "",
  apply_mask_for_mlm = FALSE,
  extras_for_debugging = FALSE
)
}
\arguments{
\item{doc}{file path, url, or raw vector to document (currently pdf only)}

\item{tokenizer}{tokenizer function to apply to words extracted from image. Currently,
{hftokenizers}, {tokenizer.bpe} and {sentencepiece} tokenizer are supported.}

\item{add_batch_dim}{(boolean) add a extra dimension to tensor for batch encoding}

\item{target_geometry}{image target magik geometry expected by the image model input}

\item{max_seq_len}{size of the embedding vector in tokens}

\item{save_to_disk}{(boolean) shall we save the result onto disk}

\item{path_to_save}{result path}

\item{apply_mask_for_mlm}{add mask to the language model}

\item{extras_for_debugging}{additionnal feature for debugging purposes}
}
\value{
a list of named tensors
}
\description{
Turn document into docformer torch tensor input feature
}
