% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_model.R
\name{docformer_fit}
\alias{docformer_fit}
\alias{docformer_fit.default}
\alias{docformer_fit.docformer_tensor}
\title{Docformer model}
\usage{
docformer_fit(x, ...)

\method{docformer_fit}{default}(x, ...)

\method{docformer_fit}{docformer_tensor}(x, config = docformer_config(), ...)
}
\arguments{
\item{x}{a \code{docformer_tensor} list, as a result of the \verb{create_features_from_*} functions}

\item{...}{Model hyperparameters.
Any hyperparameters set here will update those set by the config argument.
See \code{\link[=docformer_config]{docformer_config()}} for a list of all possible hyperparameters.}

\item{config}{A set of hyperparameters created using the \code{docformer_config} function.
If no argument is supplied, this will use the default values in \code{\link[=docformer_config]{docformer_config()}}.}

\item{y}{a \strong{data frame}}

\item{docformer_model}{A previously fitted DocFormer model object to continue the fitting on.
if \code{NULL} (the default) a brand new model is initialized.}

\item{from_epoch}{When a \code{docformer_model} is provided, restore the network weights from a specific epoch.
Default is last available checkpoint for restored model, or last epoch for in-memory model.}
}
\value{
A DocFormer model object of class \code{docformer_fit} It can be used for serialization, predictions, or further fitting.
}
\description{
Fits the \href{https://arxiv.org/abs/2106.11539}{DocFormer: End-to-End Transformer for Document Understanding} model
}
\section{Fitting a pre-trained model}{


When providing a parent \code{docformer_model} parameter, the model fitting resumes from that model weights
at the following epoch:
\itemize{
\item last fitted epoch for a model already in torch context
\item Last model checkpoint epoch for a model loaded from file
\item the epoch related to a checkpoint matching or preceding the \code{from_epoch} value if provided
The model fitting metrics append on top of the parent metrics in the returned TabNet model.
}
}

