% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_model.R
\name{docformer_pretrain}
\alias{docformer_pretrain}
\title{Docformer Self-supervised training task}
\usage{
docformer_pretrain(
  config,
  train_dataloader,
  val_dataloader,
  device,
  epochs,
  path,
  classes,
  lr = 5e-05,
  weights = weights
)
}
\arguments{
\item{weights}{}
}
\value{

}
\description{
Docformer Self-supervised training task
}
