% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutlm_network.R
\name{LayoutLMModel}
\alias{LayoutLMModel}
\title{The LayoutLM model was proposed in \href{https://arxiv.org/abs/1912.13318}{LayoutLM: Pre-training of Text and Layout for Document ImageUnderstanding}
by Yiheng Xu, Minghao Li, Lei Cui, Shaohan Huang, Furu Wei and Ming Zhou.}
\usage{
LayoutLMModel(config)
}
\description{
This model is a torch \href{https://torch.mlverse.org/docs/reference/nn_module.html}{nn_module()}. Use
it as a regular module and refer to the documentation for all matter related to general usage and
behavior.
}
\details{
Parameters:
config (\code{\link{LayoutLMConfig}}): Model configuration class with all the parameters of the model.
Initializing with a config file does not load the weights associated with the model, only the
configuration. Check out the \code{\link{~PreTrainedModel$from_pretrained}} method to load the model weights.
}
