% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{save_featureRDS}
\alias{save_featureRDS}
\title{Save feature tensor to disk}
\usage{
save_featureRDS(docformer_tensor, file)
}
\arguments{
\item{docformer_tensor}{: the \code{docformer_tensor} tensor list to save}

\item{file}{: destination file}
}
\description{
Save feature tensor to disk
}
